evilpanel/requirements.txt
# EvilPanel Python Dependencies
# Source: plan.md Section 6.1

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Proxy & Networking
mitmproxy>=10.2.0
httpx>=0.26.0
aiohttp>=3.9.0

# Security & Crypto
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# YAML & Config
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0
email-validator>=2.1.0

# WebSocket
websockets>=12.0

# Browser Automation (session injection)
playwright>=1.41.0
selenium>=4.17.0

# Notifications
python-telegram-bot>=20.7

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

